Creates a comparison table for miRTarBase to be used for custom integration

This function downloads the miRTarBase database for the organism of choice, filters it according to user-specified values and formats ready for custom integration in generateShinyApp.

Usage

preprocess_miRTarBase(
  download.dir = ".",
  download.method = "auto",
  mirtarbase.file = NULL,
  organism.code,
  org.db,
  support.type = c(),
  validation.method = c(),
  reference = c("mRNA", "miRNA"),
  print.support.types = FALSE,
  print.validation.methods = FALSE
)

Arguments

download.dir

Directory where miRTarBase database will be downloaded.

download.method

Method for downloading miRTarBase file through download.file, see download.file documentation for options for your operating system.

mirtarbase.file

Path to pre-downloaded miRTarBase file for your organism. If this is left NULL then the file will be downloaded.

organism.code

Three letter code for the organism of choice. See miRTarBase website for options. For human, enter 'hsa' and for mouse, 'mmu'.

org.db

database for annotations to transform ENSEMBL IDs to gene names; a list of bioconductor packaged databases can be found with BiocManager::available("^org\.").

support.type

Subset of entries of the 'Support Type' field in miRTarBase. Only these values will be kept. To find the options available for your organism of choice, run the function once with print.support.types = TRUE.

validation.method

Subset of entries of 'Experiments' field in miRTarBase. Only these values will be kept. To find the options available for your organism of choice, run the function once with print.validation.methods = TRUE.

reference

Should the reference category be mRNA or miRNA? The reference category chosen here must match the reference category chosen in custom.integration in generateShinyApp. Default in mRNA.

print.support.types, print.validation.methods

Should options for Support Type and Experiments be displayed? Default is FALSE.

Value

A dataframe with Reference_ID/Name and Comparison_ID/Name columns which can be supplied to custom.integration in generateShinyApp

Examples

comparison.table <- preprocess_miRTarBase(
  mirtarbase.file = system.file("extdata", "mmu_MTI_sub.xls", package = "bulkAnalyseR"),
  organism.code = "mmu",
  org.db = "org.Mm.eg.db",
  support.type = "Functional MTI",
  validation.method = "Luciferase reporter assay",
  reference = "miRNA")